requirements.txt
cryptography
fire
install_playwright
jsonschema
numpy<=2.2.6
pandas
pillow
playwright==1.54.0
pydantic>=2.10.6
requests
retry
pytz
strands-agents>=0.1.9
strands-agents-tools
Deprecated
websockets
nova-act
python-dotenv
boto3
